-organized files into subpackages as per android best practices
