+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/alidd/AndroidStudioProjects/CST-483-Project-3-Mobile/app/src/main/res/layout/activity_landing.xml" value="0.32760416666666664" />
         <entry key="..\:/Users/alidd/AndroidStudioProjects/CST-483-Project-3-Mobile/app/src/main/res/layout/fragment_student_home.xml" value="0.5744791666666667" />
         <entry key="..\:/Users/orozc/AndroidStudioProjects/Project3Mobile/app/src/main/res/layout/activity_main.xml" value="0.2753623188405797" />
         <entry key="..\:/Users/orozc/AndroidStudioProjects/Project3Mobile/app/src/main/res/layout/fragment_landing.xml" value="0.29322916666666665" />
