+++ app/src/main/java/com/example/project3mobile/activities/LandingActivity.java
@@ -1,8 +1,10 @@
-package com.example.project3mobile;
+package com.example.project3mobile.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.project3mobile.R;
 
 public class LandingActivity extends AppCompatActivity {
