+++ app/src/main/java/com/example/project3mobile/fragments/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.project3mobile;
+package com.example.project3mobile.fragments;
 
 import android.os.Bundle;
 
@@ -8,23 +8,25 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.project3mobile.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link StudentHomeFragment#newInstance} factory method to
+ * Use the {@link LoginFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class StudentHomeFragment extends Fragment {
+public class LoginFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
+    private static final String mUsername = "username";
+    private static final String mPassword = "password";
 
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
 
-    public StudentHomeFragment() {
+    public LoginFragment() {
         // Required empty public constructor
     }
 
@@ -34,14 +36,14 @@ public class StudentHomeFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment StudentHomeFragment.
+     * @return A new instance of fragment LoginFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static StudentHomeFragment newInstance(String param1, String param2) {
-        StudentHomeFragment fragment = new StudentHomeFragment();
+    public static LoginFragment newInstance(String param1, String param2) {
+        LoginFragment fragment = new LoginFragment();
         Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
+        args.putString(mUsername, param1);
+        args.putString(mPassword, param2);
         fragment.setArguments(args);
         return fragment;
     }
@@ -50,8 +52,8 @@ public class StudentHomeFragment extends Fragment {
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
+            mParam1 = getArguments().getString(mUsername);
+            mParam2 = getArguments().getString(mPassword);
         }
     }
 
@@ -59,6 +61,6 @@ public class StudentHomeFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_student_home, container, false);
+        return inflater.inflate(R.layout.fragment_login, container, false);
     }
 }
