+++ app/src/main/java/com/example/project3mobile/fragments/RegisterFragment.java
@@ -1,4 +1,4 @@
-package com.example.project3mobile;
+package com.example.project3mobile.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.project3mobile.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link LandingFragment#newInstance} factory method to
+ * Use the {@link RegisterFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class LandingFragment extends Fragment {
+public class RegisterFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class LandingFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public LandingFragment() {
+    public RegisterFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class LandingFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment LandingFragment.
+     * @return A new instance of fragment RegisterFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static LandingFragment newInstance(String param1, String param2) {
-        LandingFragment fragment = new LandingFragment();
+    public static RegisterFragment newInstance(String param1, String param2) {
+        RegisterFragment fragment = new RegisterFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class LandingFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_landing, container, false);
+        return inflater.inflate(R.layout.fragment_register, container, false);
     }
 }
